+++ EduOne/main/models.py
@@ -1,5 +1,4 @@
 from django.db import models
-from django.forms import ModelForm
 from django.contrib.auth.models import User
 from django.utils.timezone import now
 
@@ -200,24 +199,6 @@ class Event(models.Model):
     def __str__(self):
         return "{}, Date: {}, From: {}, To: {}".format(self.eventPlanner, self.eventDate, self.timeFrom, self.timeTo)
 
-#This form is used by the views.schedule_add to add create new Event objects
-class EventForm(ModelForm):
-    class Meta:
-        model = Event
-        exclude = ['eventPlanner']
-        labels = {
-            'description': ('Event Description'),
-            'eventDate': ('Date of Event'),
-            'timeFrom': ('Start Time of Event'),
-            'timeTo': ('End Time of Event'),
-        }
-        help_texts = {
-            'description' : ('The details of the event goes here.'),
-            'eventDate' : ('The date of the event goes here.'),
-            'timeFrom' : ('The start time of the event goes here.'),
-            'timeTo' : ('The end time of the event goes here.'),
-        }
-
 
 APPOINTMENT_STATUS = (
     ('pending'),
@@ -256,22 +237,3 @@ class Announcement(models.Model):
     def __str__(self):
         return "Creator: {}, isEvent: {}".format(self.creator, self.isEvent)
 
-#This form is used by the views.schedule_add to add create new Announcement objects
-class AnnoucementForm(ModelForm):
-    class Meta:
-        model = Announcement
-        exclude = ['creator']
-        labels = {
-            'description': ('Announcement Description'),
-            'isEvent' : ('Is this announcement part of a event?'),
-            'eventDate': ('Date of Announcement'),
-            'eventTimeFrom': ('Start Time of Announcement'),
-            'eventTimeTo': ('End Time of Announcement'),
-        }
-        help_texts = {
-            'description': ('The announcement description goes here.'),
-            'isEvent' : ('Specifies whether a not the announcement is part of a event.'),
-            'eventDate': ('The date of the announcement goes here.'),
-            'eventTimeFrom': ('The time where the announcement\'s event(s) starts goes here.'),
-            'eventTimeTo': ('The time where the announcement\'s event(s) ends goes here.'),
-        }
