+++ EduOne/main/templates/users/logout.html
@@ -1,0 +1,9 @@
+{% extends "common/header_login.html" %}
+
+{% block content %}
+
+	<h2>You have successfully logout!</h2>
+	<p><a href="{% url 'login' %}"> Click here to login again! </a><p>
+
+
+{% endblock %}
