+++ EduOne/main/urls.py
@@ -15,10 +15,13 @@ Including another URLconf
 """
 
 from django.urls import path
-from . import views
+from django.contrib.auth import views as auth_views
+from . import views, forms
 
 urlpatterns = [
-    path('', views.home),
+    path('', auth_views.LoginView.as_view(template_name='users/login.html'), name='login'),
+    path('logout/', auth_views.LogoutView.as_view(template_name='users/logout.html'), name='logout'),
+    path('home/', views.home, name='main-home'),
     path('schedule/', views.schedule, name='schedule-home' ),
     path('schedule/add/', views.schedule_add, name='schedule-add'),
     path('schedule/add=<slug:current>/', views.schedule_add, name='schedule-add'),
