+++ EduOne/main/views.py
@@ -1,20 +1,22 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
-from .models import ReportCardPage, SubjectGrade, EventForm, AnnoucementForm
+from .models import ReportCardPage, SubjectGrade
+from .forms import EventForm, AnnoucementForm
 
 # Create your views here.
+@login_required
 def home(request):
     return render(request, 'index.html', {'active_page': 'home'})
 
 
 #This function is for the main page of scheduling, can be used by both teachers and parents
-#@login_required
+@login_required
 def schedule(request):
 	return render(request, 'schedule/schedule.html', {'active_page': 'schedule'})
 
 #This function is for the teachers to add new event that can be announcments/events
-#@login_required
+@login_required
 def schedule_add(request, current='events'):
 	event_categories = ('events', 'announcements')
 	if current not in event_categories:
@@ -24,7 +26,8 @@ def schedule_add(request, current='events'):
 		return redirect ('schedule-add')
 	
 	elif current == event_categories[0]:
-		form = EventForm()
+		form = EventForm(request.POST)
+
 		#Create objects to add into classes!
 
 	elif current == event_categories[1]:
